CountMin_CU Soporta inserciones de kmers
Se actualizo CountMin_CU para que soporte las inserciones de kmers

+++ CountMin_CU.h
@@ -11,7 +11,7 @@ class CountMin_CU
     private:
     int d;
     int w;
-    std::vector<std::vector<int>> C;
+    std::vector<std::vector<uint64_t>> C;
 
     public:
     CountMin_CU(int d, int w);
